fix sonar issues
