+++ frontend/src/ui-components/AcmTable/AcmTable.test.tsx
@@ -8,8 +8,8 @@ import { configureAxe } from 'jest-axe'
 import { useState } from 'react'
 import { AcmDropdown } from '../AcmDropdown/AcmDropdown'
 import { AcmEmptyState } from '../AcmEmptyState'
-import { AcmTable, AcmTablePaginationContextProvider, AcmTableProps, ExportableIRow } from './AcmTable'
-import { ITableAdvancedFilter } from './AcmTableToolbar'
+import { AcmTable, AcmTablePaginationContextProvider } from './AcmTable'
+import { AcmTableProps, ExportableIRow, ITableAdvancedFilter } from './AcmTableTypes'
 
 import { exampleData } from './AcmTable.stories'
 import { MemoryRouter, Route, Routes } from 'react-router-dom-v5-compat'
