+++ frontend/src/ui-components/index.ts
@@ -37,6 +37,7 @@ export * from './AcmSelect/AcmSelect'
 export * from './AcmSummaryList/AcmSummaryList'
 export * from './AcmTable/AcmTable'
 export * from './AcmTable/AcmTableToolbar'
+export * from './AcmTable/AcmTableTypes'
 export * from './AcmTextArea/AcmTextArea'
 export * from './AcmTextInput/AcmTextInput'
 export * from './AcmTile/AcmTile'
